Add interactive install integration test
Most of the unhappy paths are already handled by the UI, so there's
really no way installing the service could fail in practice.

Initially I wanted to do a table test for this as well, but then I
realized that due to the strictness of the UI it was impossible to
trigger an install failure (e.g. due to bad config path or dirs).