Add system proxy policy manager
For now we only handle system-wide proxy settings, and enforce them
via a D-Bus call to our ubuntu-proxy-manager service.

Prior to making any D-Bus calls, we check if the ubuntu-proxy-manager is
installed. If the GPO contains proxy entries and the package is not
installed, we fail hard. Otherwise, we just silently return.

The manager returns any error returned by the D-Bus call.

Closes DEENG-554