Add integration tests for system proxy manager
This change fully mocks the com.ubuntu.ProxyManager.Apply call.

In the default case, the method is mocked to pass.

When 'apply_proxy_fail' is set, the call will reach the Apply method and
fail.

When 'no_proxy_object' is set, the object won't be mocked at all, and
the call will fail. This simulates the behavior when the
ubuntu-proxy-manager package is not installed.