Add short sleep after VM creation
Maybe it's because the GitHub Actions runners run on Azure itself so
speeds are blazing fast, but in some cases Azure reports the VM as
created before the first-run script finishes running, meaning the SSH
keys are not yet properly set up.

The simplest way to mitigate this is to wait for a bit before creating
the SSH connection. If this proves problematic we can revisit and
perform the SSH configuration changes when preparing the base VM
template as opposed to on the first run of the VM.